tools/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ectf25"
version = "2025.01.15"
requires-python = ">=3.11"
dependencies = [
    "loguru",
    "pyserial",
    "tqdm",
    "pycrypto",
    "pycryptodome",
    "cryptography",
]

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
